PLAT-228: Catch Replication failure on litecore
The replicator object is failing after 6 minutes of idle time. This is an internal issue with
couchbase and for timesake this is caught to restart the replicator. This will only catch the error that states "Remote Error" and will not restart if some other issue occurs. This is to avoid it looping on a valid failure.