Refactor observation patter with Nimbus
Made some changes to Abstract Observer Class:

Now implements 2 callbacks - DocumentChangeCallback and SyncStatusCallback

DocumentChangeCallback - Callback that will be called when there a document sync (from cloud)
has been detected. This will be called for every document that changed. The argument is the body of the document.

SyncStatusCallback - Callback that will be called when there is a change in state in Nimbus. The argument is the NimbusSyncInfo struct.