Update path and argument for HCS
The working directory changes depending if you are starting the
usb-pd.exe with qt or manually calling the exe. This was only affecting
the HCS call and the path to its config file. Update the working
directory of the Qprocess so that it matches the app directory. This
will force it to be the same everytime.

Also, update the arguments to point to the new hcs config file.