Change fleece::alloc_slice refereces to FLSlice
Windows could not link to alloc_slice in fleece namespace. This somehow works on windows and MacOS butWindow cross-compiles fail with a error to references. From what can be seen, the windows DLLs do NOT have references to the alloc_slice structure and therefore need to use the FLSLICE that is available publicly.