PLAT-299: Accept commands via JSON
New api function Command(). This will be a REST api and accepts json. Ideally weuse a schema validator but currently rapidjson is giving errors.

This command will remove the need for discrete read(), write() and other functions.

Currently the only command supported is "connect_documents" : "all" and is manually checked (ugh).

The schema file is loaded but not used.