PLAT-380: Launch HCS with Spyglass for Mac/Linux
We have seen timing issues that occur when we launch HCS with spyglass. This is currently only available through Windows. Expand this functionality so Mac and Linux can launch as well to help with troubleshooting.

This functionality will require START_SERVICES to be defined. It is recommened that a new build configuration be made ("Rel + Services") to specifically define this at build time.

To do this under QT: Go to Projects Page-> Build Steps-> Make Details-> Make arguments-> Add "DEFINES+="-DSTART_SERVICES"