PLAT-455: Redesign Board and Navigation QML
Preliminary framework for the navigation control. This is barebone skeleton to start and will be incrementally revisioned.

This framework is state machine controlled by the navigationControl javascript. All events must go throu the navigationControl and it will determine what will be displayed. All files are currently placeholders for the real files to be linked later.