PLAT-557: Catch errors on auto select
The destruction of dynamic views throws an error that much be caught in navigation control. If not caught the Welcome screen will fail to load its platform list do to the thrown exception. Properly handle the error and also make the auto select feature a boolean to enable / disable easily.