Platform selector: dynamic selector
Updated HCS to tell coreInterface if the platform list is a connected platform list or a list of all platforms. Updated coreInterface to differentiate between the two as well as the JS in the UI dealing with these lists. Updated platform_selection.js to update the all-platforms list with the info from the connected-platforms list. Added logic to use hardcoded platform list in lieu of list supplied by HCS.