Merged in SCT-585-hcs-does-not-send-empty-connecte (pull request #829)
HCS: bugfix for sending platform list populated with disconnected platforms

On disconnection, BoardsController dispatches a command to HCS to send platform
list to UI, then clears the connected platforms list. These occur asynchronously,
which allowed the list to be created before it was cleared in rare cases. Added a
disconnected state to boards since the list creation checks board state.

Approved-by: Luay Alshawi <luay.alshawi@onsemi.com>