sds: let HCS exit with non-standard special code to indicate desired quit
When we start the HCS multiple times only first one stay running; the next instances quit with special exit code.
SDS has to watch for it and it should not disconnect...
Note: this need to be fixed; because HCS indicate UP and finished second HCS
may indicate QUIT (and  in this order of events it make troubles).