ResourceLoader: Shortened async wait time
Realized that it would be better to not unregister the resource than
it would be to wait too long and not delete it if the app closes or
something. Since the logic is to open any available local RCC,
this could get into a stuck state where old RCC loads and cannot update.