cmake: use proper order of project submodules
CMake is working on 2 phases - configuration & project files generation.
The first one is sequential i.e. the CMake option may be used after it
is defined.
In our case particular components depends if particular apps or features
are enabled (to speedup build).
Using this speedup optimization for 'extern' project would be possible
but it will require more changes.

This is regression; it was meantime fixed by commit 797404e.